add basic math functions
