+++ docs/src/piccolo/functions/index.rst
@@ -8,5 +8,6 @@ Functions can be used to modify how queries are run, and what is returned.
 
     ./basic_usage
     ./string
+    ./math
     ./type_conversion
     ./aggregate
