+++ docs/src/piccolo/functions/math.rst
@@ -1,0 +1,28 @@
+Math functions
+==============
+
+.. currentmodule:: piccolo.query.functions.math
+
+Abs
+---
+
+.. autoclass:: Abs
+    :class-doc-from: class
+
+Ceil
+----
+
+.. autoclass:: Ceil
+    :class-doc-from: class
+
+Floor
+-----
+
+.. autoclass:: Floor
+    :class-doc-from: class
+
+Round
+-----
+
+.. autoclass:: Round
+    :class-doc-from: class
