+++ piccolo/query/functions/__init__.py
@@ -1,17 +1,22 @@
 from .aggregate import Avg, Count, Max, Min, Sum
+from .math import Abs, Ceil, Floor, Round
 from .string import Length, Lower, Ltrim, Reverse, Rtrim, Upper
 from .type_conversion import Cast
 
 __all__ = (
+    "Abs",
     "Avg",
     "Cast",
+    "Ceil",
     "Count",
+    "Floor",
     "Length",
     "Lower",
     "Ltrim",
     "Max",
     "Min",
     "Reverse",
+    "Round",
     "Rtrim",
     "Sum",
     "Upper",
